fixed #95
the assessement persist operation were using the original
kassenzeichenBean instead of the persisted one.

the kassenzeichenBean is persisted a first time for
the save operation, and than the assessementBean is
created with a reference to the kassenzeichenBean.
If the kassenzeichenBean is used instead of the
persisted kassenzeichenBean, then the new metaobjects
in the arrays are still stated as "NEW" which is the
cause for the double objects