issue: #98
NumberStringComparator used for the first column of the correct-Table.
This ensures that the sortorder algorithm is the same as of the main table.

Instead of applying the corrections to the model of the main table,
the changes are now applyed to the model of the correct table. this
makes the correct table totaly independent to the main tables column orders.