fixed #111
the layout had to be load AFTER the main frame become visible
and to be stored BEFORE it is disposed.

using now the isInit flag to check is the application is in
init phase and to prevent to load the layout to early.

loadLayoutWhenOpenedAdapter loads now the layout when the
WindowOpened event of the main frame is triggerd.